AddDomainEvents+OutboxMessagePattern+FullProjectTest

+++ TransitNova.BusinessLayer/Features/Location/Countries/Handlers/ApplyCommands/DeleteCountryHandler.cs
@@ -4,6 +4,7 @@ using TransitNova.BusinessLayer.Common.ResultPattern;
 using TransitNova.BusinessLayer.Features.Location.Countries.Commands;
 using TransitNova.BusinessLayer.Interfaces.Repositories.LocationRepository;
 using TransitNova.BusinessLayer.Interfaces.Services.UnitOfWork;
+using TransitNova.Domain.Entities.MainEntities;
 
 namespace TransitNova.BusinessLayer.Features.Location.Countries.Handlers.ApplyCommands
 {
@@ -15,8 +16,17 @@ namespace TransitNova.BusinessLayer.Features.Location.Countries.Handlers.ApplyCo
     {
         public async Task<BaseResult> Handle(DeleteCountryCommand request, CancellationToken ct)
         {
-            await repository.DeleteAsync(request.Id, ct);
+            var country = await repository.GetByIdAsync<Country>(request.Id,ct);
+            if (country == null)
+                return BaseResult.NotFound(Errors.NotFound("Country Not Found"));
 
+
+            var deleted =  await repository.DeleteAsync(request.Id, ct);
+            if (!deleted)
+            {
+                logger.LogWarning("Country delete failed because government was not found. CountryId: {CountryId}", request.Id);
+                return BaseResult.UnExpected(Errors.FailedOperation("Un Error Occured While Removing CountryCountry"));
+            }
             await unitOfWork.SaveChangesAsync(ct);
             logger.LogInformation("Country deleted successfully. Id: {CountryId}", request.Id);
             return BaseResult.Success();
